Create PlayerWindowController.swift
Refactoring: make `PlayerWindowController` the super class of
`MainWindowController` and `MiniPlayerWindowController`,
merge and reuse similar code (w/ IBOutlet, IBAction which are automatically
matched by Xcode)
